Add User Logging Class
The user_logger is used to write logs as usual prepended with 'USER - '
but also send the log message over the websocket.

The connection manager has been moved out of the api folder and put in
it's own seperate web socket folder. This was to fix an issue with a
circular dependency caused by calling the user_logger in the agents
due to the __init__ file in api.

<body>