Rename user_logger and add prefix argument
Extends the existing user logger functionality to be able to prefix logs
with an enum. The only current value is USER but this could be extended
as needed.
As we've now extended the user_logger to be ready to handle more than
just logs prefixed with user it has been renamed to log_publisher